Move code back to a single file
I think it is better to have all of my up to date code in a single
file instead of creating a new file every single time I do anything
different. This will keep the repository much cleaner also. Old
files can always be searched for in the git history.