Add code to render texture on square
Texture is loaded by a library and sent to GL and all that fancy
stuff. It is very confusing but I think I might get most of it.

Also it said to check the shader logs... I should look up how to do
that for future reference because I honestly have no idea what the
shader logs are or where they are located.