Add code to load vertices to model.h
The code is UNTESTED! It will first check for a .vrt file,
if found will compare timestamps with the .obj file. If .vrt is older,
it will generate a new .vrt by calling convertWavefront from wfc.c.

If a .vrt file is not found, it will also generate it. If a .vrt is
found but a .obj is not found, it just reads the .vrt and returns.

I was thinking about it erroring out if a .obj was not found but I
decided against it.