Possibly fix memory heap issues
When allocating memory for a string that will be initialized with
strcpy, you MUST allocate malloc(strlen(other) + 1) to make room
for the '\0' at the end!! Other wise there could be possible issues.

This could be what was causing the issues.

In other news, valgrind reports that there are no memory issues
in this program now so HOORAY.