Read spec map name from .mat, store in .vrt
If a spec map is not found in the .mat, the program will search for
matname_spec.png and store that in the .vrt. If even that is not found,
it will write 0 for the size of the spec_map name, print an error, and
continue on with it's life because it really does not care.